+++ ccres_disdrometer_processing/processing/preprocessed_file2processed_weather.py
@@ -279,7 +279,9 @@ def compute_todays_events_stats_weather(ds, Ze_ds, conf, qc_ds, start, end):
     # n is the number of events to store in the dataset
     # i.e. the number of events which begin at day D
 
-    stats_ds = xr.Dataset(coords=dict(events=(["events"], np.linspace(1, n, n))))
+    stats_ds = xr.Dataset(
+        coords=dict(events=(["events"], np.linspace(1, n, n, dtype="int32")))
+    )
 
     dZ_mean, dZ_med, dZ_q1, dZ_q3, dZ_min, dZ_max = (
         np.zeros(n),
@@ -405,7 +407,7 @@ def compute_todays_events_stats_weather(ds, Ze_ds, conf, qc_ds, start, end):
         data=end_event, dims=["events"], attrs={"long_name": "event end epoch"}
     )
     stats_ds["event_length"] = xr.DataArray(
-        data=event_length,
+        data=event_length.astype("i4"),
         dims=["events"],
         attrs={"long_name": "event duration", "unit": "mn"},
     )
@@ -438,7 +440,7 @@ def compute_todays_events_stats_weather(ds, Ze_ds, conf, qc_ds, start, end):
     )
 
     stats_ds["nb_dz_computable_pts"] = xr.DataArray(
-        data=nb_dz_computable_pts,
+        data=nb_dz_computable_pts.astype("i4"),
         dims=["events"],
         attrs={
             "long_name": "number of timesteps for which Delta Z can be computed",
@@ -504,7 +506,7 @@ def compute_todays_events_stats_weather(ds, Ze_ds, conf, qc_ds, start, end):
     )
 
     stats_ds["good_points_number"] = xr.DataArray(
-        data=nb_good_points,
+        data=nb_good_points.astype("i4"),
         dims=["events"],
         attrs={
             "long_name": "number of timesteps where all checks are good",
